Added non-root user and git-email

+++ dev-sesh.sh
@@ -5,32 +5,27 @@
 IMAGE_NAME="dev-container:latest"
 CONTAINER_NAME="dev-container-instance"
 SOURCE_CODE_PATH="/home/rahulrajesh/tee-project/src"
-
-start_docker_daemon() {
-	echo "Starting Docker daemon..."
-	sudo systemctl start docker
-}
-
-stop_docker_daemon() {
-	echo "Stopping Docker daemon..."
-	sudo systemctl stop docker.socket;
-	sudo systemctl stop docker.service;
-}
+TARGET_PATH="/home/user/"
 
 bring_up_container() {
-	echo "Starting a new container..."
-	docker run --rm -it \
-		--name $CONTAINER_NAME \
-		--mount type=bind,source=/dev,target=/dev \
-		--mount type=bind,source=${SOURCE_CODE_PATH},target=/workspace \
-		--device-cgroup-rule "c 188:* rwm" \
-		--device-cgroup-rule "c 189:* rwm" \
-		$IMAGE_NAME
+    echo "Checking for existing container..."
+    if docker ps -a --format '{{.Names}}' | grep -q "^${CONTAINER_NAME}$"; then
+        echo "Container ${CONTAINER_NAME} already exists. Creating new shell in existing container..."
+        docker exec -it ${CONTAINER_NAME} bash
+    else
+        echo "Starting a new container..."
+        docker run --rm -it \
+            --name ${CONTAINER_NAME} \
+            --mount type=bind,source=/dev,target=/dev \
+            --mount type=bind,source="${SOURCE_CODE_PATH}",target="${TARGET_PATH}" \
+            --device-cgroup-rule "c 188:* rwm" \
+            --device-cgroup-rule "c 189:* rwm" \
+            ${IMAGE_NAME}
+    fi
 }
 
 case "$1" in
 	up)
-		start_docker_daemon
 		bring_up_container
 		;;
 	*)
